+++ pkg/dev_compiler/tool/input_sdk/private/ddc_runtime/generators.dart
@@ -11,8 +11,9 @@
 /// stepping stone for proposed ES7 async/await, and uses ES6 Promises.
 part of dart._runtime;
 
-// TODO(vsm): Remove once this flag is the default.
-bool startAsyncSynchronously = false;
+// TODO(vsm): Remove once this flag we've removed the ability to
+// whitelist / fallback on the old behavior.
+bool startAsyncSynchronously = true;
 void setStartAsyncSynchronously([bool value = true]) {
   startAsyncSynchronously = value;
 }