+++ tools/testing/dart/test_suite.dart
@@ -1062,7 +1062,7 @@ class StandardTestSuite extends TestSuite {
         // Always run with synchronous starts of `async` functions.
         // If we want to make this dependent on other parameters or flags,
         // this flag could be become conditional.
-        content = dartdevcHtml(nameNoExt, jsDir, buildDir, syncAsync: true);
+        content = dartdevcHtml(nameNoExt, jsDir, buildDir);
       }
     }